app LastIndex types: make LastIndexedReducerState.lastIndexed nullable
This seems to be the expected behavior based on most usage sites (and the
defaults in the parser), but there may be a few users that need to be updated
to handle nullness. I think semantically null is different from 0 in this case
(no update vs update at 1970), so I think it's correct to change the type and
fix others later.

Also add a parse-time payload undefined check.

Signed-off-by: Dorian Johnson <2020@dorianj.net>