Expose config for mTLS certificates.
A client certificate can be specified to present with requests to
metadata and search services for mTLS. The configuration is given
as a pair of paths to the PEM encoded cert and key files. This is
described in more detail in the requests documentation:
https://requests.readthedocs.io/en/master/user/advanced/#client-side-certificates

This assumes that the frontend should present the same client
cert to both the search and metadata services, which is consistent
with the idea that a client cert should uniquely identify the
service that is presenting it.

If the environment variables are not set, then the clients are
created as normal without any expectation of mTLS.

Signed-off-by: Dave Cameron <dcameron@digitalocean.com>