app: fix two minor type problems related to search filters
First, getFilterConfigByResource is expected to sometimes return undefined,
so mark the signature as such. This fixes a test which mocked out an undefined
return value. TBH I'm not sure how the original implmentation passed the type
checker, I think there's an `any` in the chain that defeated it.

Second, in the call to `connect`, don't specify null for optional props, it
confuses the typechecker. Instead just omit them.

Signed-off-by: Dorian Johnson <2020@dorianj.net>