app tableMetadata duck: minor fix in generateOwnerUpdateRequests
Build the return via map, not via forEach / push, removes LOC and obviates
needing to explictly type it.

Also add the any type to the call signature. Previously this was implied due to
the opaque type of the return array builder. But as the nearby TODO indicates,
the callsites freak out when they learn the return type is
`Promise<AxiosResponse<any> | undefined>>[]`, and I don't want to dig deeper
into that.

Signed-off-by: Dorian Johnson <2020@dorianj.net>