app search ducks test: specify default SearchType in
These types are pretty unhelpful, in that there are a bunch of keys that
aren't actually used for certain search types, and are thus ignored in most
practices. This one is no different than the others, so just specify a random
default.

I'm not 100% certain in this change, but I can't find anywhere that branches
on searchType being null, so I think it's safe.

Signed-off-by: Dorian Johnson <2020@dorianj.net>