app config tests: <any> blackhole two tests
These tests dig deep into the config tree and pull out specific combinations
of keys from dicts the author "knows" exist. That's fine, but it does not
please the type checker.

Instead, blackhole the reference into <any> and add a runtime defininition
check, that way if the config structure changes it'll be flagged as a test
failure.

Signed-off-by: Dorian Johnson <2020@dorianj.net>